SRC/APPLICATION-SOURCE-CODE/requirements.txt
flask
wikipedia
mysql-connector-python